move local_account module into testing package
